Fix compile errors

+++ RFW/RenderSystem/utils/src/utils/RandomGenerator.h
@@ -26,23 +26,26 @@ class RandomGenerator
 	{
 		if (fabs(N.x) > fabs(N.y))
 		{
-			Nt = glm::vec3(N.z, 0.f, -N.x) / sqrt(N.x * N.x + N.z * N.z);
+			Nt = glm::vec3(N.z, 0.f, -N.x) / static_cast<float>(sqrt(N.x * N.x + N.z * N.z));
 		}
 		else
 		{
-			Nt = glm::vec3(0.f, -N.z, N.y) / sqrt(N.y * N.y + N.z * N.z);
+			Nt = glm::vec3(0.f, -N.z, N.y) / static_cast<float>(sqrt(N.y * N.y + N.z * N.z));
 		}
 
 		Nb = cross(N, Nt);
 	}
 
-	inline glm::vec3 localToWorld(const glm::vec3 &sample, const glm::vec3 &Nt, const glm::vec3 &Nb, const glm::vec3 &normal)
+	inline glm::vec3 localToWorld(const glm::vec3 &sample, const glm::vec3 &Nt, const glm::vec3 &Nb,
+								  const glm::vec3 &normal)
 	{
-		return glm::vec3(sample.x * Nb.x + sample.y * normal.x + sample.z * Nt.x, sample.x * Nb.y + sample.y * normal.y + sample.z * Nt.y,
+		return glm::vec3(sample.x * Nb.x + sample.y * normal.x + sample.z * Nt.x,
+						 sample.x * Nb.y + sample.y * normal.y + sample.z * Nt.y,
 						 sample.x * Nb.z + sample.y * normal.z + sample.z * Nt.z);
 	}
 
-	inline glm::vec3 worldToLocalMicro(const glm::vec3 &vec, const glm::vec3 &rDirection, glm::vec3 &u, glm::vec3 &v, glm::vec3 &w)
+	inline glm::vec3 worldToLocalMicro(const glm::vec3 &vec, const glm::vec3 &rDirection, glm::vec3 &u, glm::vec3 &v,
+									   glm::vec3 &w)
 	{
 		w = vec;
 		u = glm::normalize(glm::cross(fabs(vec.x) > fabs(vec.y) ? glm::vec3(0, 1, 0) : glm::vec3(1, 0, 0), w));
@@ -51,7 +54,8 @@ class RandomGenerator
 		return glm::normalize(glm::vec3(glm::dot(u, wi), glm::dot(v, wi), glm::dot(w, wi)));
 	}
 
-	inline glm::vec3 localToWorldMicro(const glm::vec3 &wmLocal, const glm::vec3 &u, const glm::vec3 &v, const glm::vec3 &w)
+	inline glm::vec3 localToWorldMicro(const glm::vec3 &wmLocal, const glm::vec3 &u, const glm::vec3 &v,
+									   const glm::vec3 &w)
 	{
 		return u * wmLocal.x + v * wmLocal.y + w * wmLocal.z;
 	}
@@ -66,4 +70,4 @@ class RandomGenerator
 		return localToWorld(sample, Nt, Nb, normal);
 	}
 };
-} // namespace rwf::utils
+} // namespace rfw::utils
